pyproject.toml
[project]
name = "syllabi-analizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-openai>=0.0.5",
    "pypdf>=3.17.0",
    "docx2txt>=0.8",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "unstructured>=0.10.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
]
